refact: extrair chat hooks para module/hooks/chat-hooks.mjs
